\name{block.design}
\alias{block.design}
\title{Block design from given incidence matrix}
\description{This function generates the block contents from a given incidence matrix}
\usage{block.design(N)}
\arguments{
	\item{N}{incidence matrix}
	}
\value{
	\item{design}{A matrix with number of rows equal to the number of blocks and number of columns 
equal to block size. Constant block size is assumed. Treatments are numbered as 1, 2, ..., v}
       }
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{incidence matrix}
\keyword{incomplete block design}
\keyword{internal}
