\name{check.group}
\alias{check.group}
\title{g function}
\description{This function checks group}
\usage{check.group(m,n,x,y)}
\arguments{
	\item{m}{number of rows}
	\item{n}{number of columns}	
	\item{x}{x}
	\item{y}{y}	
	}
\value{
	\item{group}{same group or different group}	
  }
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}
