\name{g}
\alias{g}
\title{g function}
\description{This function computes the value of g for specified parameters.}
\usage{g(v,b,k,x,z,alpha,rho=0)}
\arguments{
	\item{v}{number of treatments}
	\item{b}{number of blocks}
	\item{k}{block size}
	\item{x}{x}
	\item{z}{z}
	\item{alpha}{alpha}
	\item{rho}{rho, default is 0}
	}
\value{
	\item{g}{value of the function}	
  }
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}
