\name{gbbpb}
\alias{gbbpb}
\title{g function}
\description{This function computes the value of g for specified parameters.}
\usage{gbbpb(v1,v2,b,k,x,z)}
\arguments{
	\item{v1}{number of test treatments}
	\item{v2}{number of control treatments}
	\item{b}{number of blocks}
	\item{k}{block size}
	\item{x}{x}
	\item{z}{z}	
	}
\value{
	\item{value}{value of the function}	
  }
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}
