\name{getr0}
\alias{getr0}
\title{getr0}
\description{This function returns the value of r0 for specified parameters.}
\usage{getr0(v,b,k)}
\arguments{
	\item{v}{number of test treatments}
	\item{b}{number of blocks}
	\item{k}{block size}	
	}
\value{
	\item{r0star}{value of r0star}	
	\item{M}{value of M}
  }
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}
