\name{getts}
\alias{getts}
\title{Value of t and s which minimize g function}
\description{This function gives the value of t and s which minimize the g function.}
\usage{getts(v,b,k,alpha,rho)}
\arguments{
	\item{v}{number of treatments}
	\item{b}{number of blocks}
	\item{k}{block size}
	\item{alpha}{weight}
	\item{rho}{rho is zero throughout.}	
	}
\value{return the values of t and s}
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}