\name{getwq}
\alias{getwq}
\title{Value of w and q which minimize g function}
\description{This function gives the value of w and q which minimize the g function.}
\usage{getwq(v1,v2,b,k)}
\arguments{
	\item{v1}{number of test treatments}
	\item{v2}{number of controls}
	\item{b}{number of blocks}
	\item{k}{block size}	
	}
\value{return the values of w and q}
\author{Baidya Nath Mandal <mandal.stat@gmail.com>}
\keyword{internal}